backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0

# Data processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# ML dependencies (already in main requirements)
scikit-learn==1.3.2
joblib==1.3.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1
